refac(circom): move `Cell` to `zkey/cell.h`
